+++ src/main/java/com/tradenity/sdk/services/CollectionService.java
@@ -5,7 +5,7 @@ import com.tradenity.sdk.client.TradenityClient;
 import com.tradenity.sdk.model.Collection;
 import com.tradenity.sdk.model.Page;
 import com.tradenity.sdk.model.PageRequest;
-import com.tradenity.sdk.model.ResourcePage;
+import com.tradenity.sdk.resources.ResourcePage;
 import com.tradenity.sdk.resources.CollectionResource;
 import retrofit2.Call;
 
@@ -92,13 +92,13 @@ public class CollectionService extends AbstractService{
         return createInstance(call);
     }
 
-    public boolean delete(String id){
+    public void delete(String id){
         Call<Void> call =  getCollectionResource().delete(id);
-        return isSuccessful(call);
+        run(call);
     }
 
-    public boolean delete(Collection collection){
-        return delete(collection.getId());
+    public void delete(Collection collection){
+        delete(collection.getId());
     }
 
     private Map<String, Object> toMap(Collection collection) {