+++ src/main/java/com/tradenity/sdk/services/CurrencyService.java
@@ -4,6 +4,7 @@ package com.tradenity.sdk.services;
 import com.tradenity.sdk.client.TradenityClient;
 import com.tradenity.sdk.model.*;
 import com.tradenity.sdk.resources.CurrencyResource;
+import com.tradenity.sdk.resources.ResourcePage;
 import retrofit2.Call;
 
 import java.util.HashMap;
@@ -74,13 +75,13 @@ public class CurrencyService extends AbstractService{
         return createInstance(call);
     }
 
-    public boolean delete(String id){
+    public void delete(String id){
         Call<Void> call =  getCurrencyResource().delete(id);
-        return isSuccessful(call);
+        run(call);
     }
 
-    public boolean delete(Currency currency){
-        return delete(currency.getId());
+    public void delete(Currency currency){
+        delete(currency.getId());
     }
 
     private Map<String, Object> toMap(Currency currency) {