+++ src/main/java/com/tradenity/sdk/services/CustomerService.java
@@ -5,7 +5,7 @@ import com.tradenity.sdk.client.TradenityClient;
 import com.tradenity.sdk.model.Customer;
 import com.tradenity.sdk.model.Page;
 import com.tradenity.sdk.model.PageRequest;
-import com.tradenity.sdk.model.ResourcePage;
+import com.tradenity.sdk.resources.ResourcePage;
 import com.tradenity.sdk.resources.CustomerResource;
 import retrofit2.Call;
 
@@ -97,13 +97,13 @@ public class CustomerService extends AbstractService{
         return createInstance(call);
     }
 
-    public boolean delete(String id){
+    public void delete(String id){
         Call<Void> call = getCustomerResource().delete(id);
-        return isSuccessful(call);
+        run(call);
     }
 
-    public boolean delete(Customer customer){
-        return delete(customer.getId());
+    public void delete(Customer customer){
+        delete(customer.getId());
     }
 
     private Map<String, Object> toMap(Customer customer) {