+++ src/test/java/com/tradenity/sdk/test/integration/BrandServiceIntegrationTest.java
@@ -59,8 +59,8 @@ public class BrandServiceIntegrationTest extends BaseServiceIntegrationTest{
 
     @Test
     public void deleteBrandTest(){
-        boolean deleted = brandService.delete("brand_001");
-        assertThat(deleted, is(true));
+        brandService.delete("brand_001");
+        //assertThat(deleted, is(true));
 
         //boolean deleted2 = brandService.delete("brand_0011");
     }