+++ Changelog.md
@@ -1,5 +1,10 @@
 ## Changelog
 
+Version 0.8.1
+- Improvement: Move ResourcePage class to resources package where it belongs, it is internal use class only.
+- Improvement: Change all 'delete' methods to return void instead of boolean. It will throw exception in case of error.
+
+
 Version 0.8.0
 
 -  Initial release